+++ locals-tags.tf
@@ -3,4 +3,10 @@ locals {
     env   = var.environment
     stack = var.stack
   } : {}
+
+  extra_tags = merge(local.default_tags, var.extra_tags)
+
+  app_gateway_tags = merge(local.extra_tags, var.app_gateway_tags)
+  ip_tags          = merge(local.extra_tags, var.ip_tags)
+  nsg_tags         = merge(local.extra_tags, var.nsg_tags)
 }